+++ electron/src/frontend/components/modals/safe-storage-popover-content.tsx
