+++ electron/src/frontend/panels/browser/browser-panel.tsx
