+++ electron/src/frontend/panels/chat/components/input/button-scroll-to-bottom.tsx
