+++ electron/src/frontend/panels/chat/components/messages/chat.message-variants.tsx
