+++ electron/src/frontend/panels/chat/components/messages/chat.spinner.tsx
