+++ electron/src/frontend/panels/chat/lib/hooks/chat.use-enter-submit.ts
