+++ electron/src/frontend/panels/chat/lib/hooks/chat.use-scroll-anchor.ts
