+++ electron/src/frontend/panels/chat/lib/hooks/chat.use-streamable-text.ts
