+++ electron/src/frontend/panels/editor/components/file-tabs/action-item.tsx
