+++ electron/src/frontend/panels/editor/file-tree/file-tree.tsx
