+++ electron/src/frontend/panels/editor/file-tree/tree-view-api.tsx
