+++ electron/src/frontend/panels/planner/planner-panel.tsx
