+++ electron/src/frontend/panels/shell/shell-panel.tsx
