+++ electron/src/frontend/panels/timeline/timeline-panel.tsx
