+++ electron/tailwind.config.ts
@@ -5,11 +5,7 @@ import containerQueries from '@tailwindcss/container-queries'
 
 const config: Config = {
     darkMode: 'class',
-    content: [
-        './src/frontend/*.{js,ts,jsx,tsx,mdx}',
-        './src/frontend/components/**/*.{js,ts,jsx,tsx,mdx}',
-        './src/frontend/components/ui/*.{js,ts,jsx,tsx,mdx}',
-    ],
+    content: ['./src/frontend/**/*.{js,ts,jsx,tsx,mdx}'],
     theme: {
         container: {
             center: true,
